generate-meal-recipe/requirements.txt
--extra-index-url https://us-central1-python.pkg.dev/mealhow/mealhow-python/simple/
openai==0.28.1
asyncio==3.4.3
aiohttp==3.9.*
functions-framework==3.*
google-cloud-ndb==2.2.2
mealhow-sdk==0.2.15
